Fixed compilation problem, removed warning.
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/trunk/supertuxkart@1815 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/game_manager.cpp
@@ -74,7 +74,7 @@ void GameManager::run()
         while( 1 )
         {
             m_curr_time = SDL_GetTicks();
-            dt = m_curr_time - m_prev_time;
+            dt =(float)(m_curr_time - m_prev_time);
 
             //This avoid wasting CPU cycles
             //1000 miliseconds / 125 frames = 125 miliseconds per frame
